+++ mintpy/objects/insar_vs_gps.py
@@ -22,135 +22,6 @@ from mintpy.defaults.plot import *
 
 
 
-############################## utilities functions ##########################################
-
-def plot_insar_vs_gps_scatter(vel_file, csv_file='gps_enu2los.csv', msk_file=None, ref_gps_site=None,
-                              xname='InSAR', vlim=None, ex_gps_sites=[], display=True):
-    """Scatter plot to compare the velocities between SAR/InSAR and GPS.
-
-    Parameters: vel_file     - str, path of InSAR LOS velocity HDF5 file.
-                ref_gps_site - str, reference GNSS site name
-                csv_file     - str, path of GNSS CSV file, generated after running view.py --gps-comp
-                msk_file     - str, path of InSAR mask file.
-                xname        - str, xaxis label
-                vlim         - list of 2 float, display value range in the unit of cm/yr
-                               Default is None to grab from data
-                               If set, the range will be used to prune the SAR and GPS observations
-                ex_gps_sites - list of str, exclude GNSS sites for analysis and plotting.
-    Example:
-        from mintpy.objects.insar_vs_gps import plot_insar_vs_gps_scatter
-        csv_file = os.path.join(work_dir, 'geo/gps_enu2los.csv')
-        vel_file = os.path.join(work_dir, 'geo/geo_velocity.h5')
-        msk_file = os.path.join(work_dir, 'geo/geo_maskTempCoh.h5')
-        plot_insar_vs_gps_scatter(vel_file, ref_gps_site='CACT', csv_file=csv_file, msk_file=msk_file, vlim=[-2.5, 2])
-    """
-
-    disp_unit = 'cm/yr'
-    unit_fac = 100.
-
-    # read GPS velocity from CSV file (generated by gps.get_gps_los_obs())
-    col_names = ['Site', 'Lon', 'Lat', 'Displacement', 'Velocity']
-    num_col = len(col_names)
-    col_types = ['U10'] + ['f8'] * (num_col - 1)
-
-    print('read GPS velocity from file: {}'.format(csv_file))
-    fc = np.genfromtxt(csv_file, dtype=col_types, delimiter=',', names=True)
-    sites = fc['Site']
-    lats = fc['Lat']
-    lons = fc['Lon']
-    gps_obs = fc[col_names[-1]] * unit_fac
-
-    if ex_gps_sites:
-        ex_flag = np.array([x in ex_gps_sites for x in sites], dtype=np.bool_)
-        if np.sum(ex_flag) > 0:
-            sites = sites[~ex_flag]
-            lats = lats[~ex_flag]
-            lons = lons[~ex_flag]
-            gps_obs = gps_obs[~ex_flag]
-
-    # read InSAR velocity
-    print('read InSAR velocity from file: {}'.format(vel_file))
-    atr = readfile.read_attribute(vel_file)
-    length, width = int(atr['LENGTH']), int(atr['WIDTH'])
-    coord = ut.coordinate(atr)
-    ys, xs = coord.geo2radar(lats, lons)[:2]
-
-    msk = readfile.read(msk_file)[0] if msk_file else np.ones((length, width), dtype=np.bool_)
-
-    num_site = sites.size
-    insar_obs = np.zeros(num_site, dtype=np.float32) * np.nan
-    prog_bar = ptime.progressBar(maxValue=num_site)
-    for i in range(num_site):
-        x, y = xs[i], ys[i]
-        if (0 <= x < width) and (0 <= y < length) and msk[y, x]:
-            box = (x, y, x+1, y+1)
-            insar_obs[i] = readfile.read(vel_file, datasetName='velocity', box=box)[0] * unit_fac
-        prog_bar.update(i+1, suffix='{}/{} {}'.format(i+1, num_site, sites[i]))
-    prog_bar.close()
-
-    off_med = np.nanmedian(insar_obs - gps_obs)
-    print(f'median offset between InSAR and GPS [before common referencing]: {off_med:.2f} cm/year')
-
-    # reference site
-    if ref_gps_site:
-        print(f'referencing both InSAR and GPS data to site: {ref_gps_site}')
-        ref_ind = sites.tolist().index(ref_gps_site)
-        gps_obs -= gps_obs[ref_ind]
-        insar_obs -= insar_obs[ref_ind]
-
-    # remove NaN value
-    print('removing sites with NaN values in GPS or {}'.format(xname))
-    flag = np.multiply(~np.isnan(insar_obs), ~np.isnan(gps_obs))
-    if vlim is not None:
-        print('pruning sites with value range: {} {}'.format(vlim, disp_unit))
-        flag *= gps_obs >= vlim[0]
-        flag *= gps_obs <= vlim[1]
-        flag *= insar_obs >= vlim[0]
-        flag *= insar_obs <= vlim[1]
-
-    gps_obs = gps_obs[flag]
-    insar_obs = insar_obs[flag]
-    sites = sites[flag]
-
-    # stats
-    print('GPS   min/max: {:.2f} / {:.2f}'.format(np.nanmin(gps_obs), np.nanmax(gps_obs)))
-    print('InSAR min/max: {:.2f} / {:.2f}'.format(np.nanmin(insar_obs), np.nanmax(insar_obs)))
-
-    rmse = np.sqrt(np.sum((insar_obs - gps_obs)**2) / (gps_obs.size - 1))
-    r2 = stats.linregress(insar_obs, gps_obs)[2]
-    print('RMSE = {:.1f} cm'.format(rmse))
-    print('R^2 = {:.2f}'.format(r2))
-
-    # plot
-    if display:
-        plt.rcParams.update({'font.size': 12})
-        if vlim is None:
-            vlim = [np.min(insar_obs), np.max(insar_obs)]
-            buffer = (vlim[1] - vlim[0]) * 0.1
-            vlim = [vlim[0] - buffer, vlim[1] + buffer]
-
-        fig, ax = plt.subplots(figsize=[4, 4])
-        ax.plot((vlim[0], vlim[1]), (vlim[0], vlim[1]), 'k--')
-        ax.plot(insar_obs, gps_obs, '.', ms=15)
-
-        # axis format
-        ax.set_xlim(vlim)
-        ax.set_ylim(vlim)
-        ax.set_xlabel(f'{xname} [{disp_unit}]')
-        ax.set_ylabel(f'GNSS [{disp_unit}]')
-        ax.set_aspect('equal', 'box')
-        fig.tight_layout()
-
-        # output
-        out_fig = '{}_vs_gps_scatter.pdf'.format(xname.lower())
-        plt.savefig(out_fig, bbox_inches='tight', transparent=True, dpi=300)
-        print('save figure to file', out_fig)
-        plt.show()
-
-    return sites, insar_obs, gps_obs
-
-
-
 ############################## beginning of insar_vs_gps class ##############################
 class insar_vs_gps:
     """ Comparing InSAR time-series with GPS time-series in LOS direction
