+++ mintpy/utils/plot.py
@@ -17,6 +17,7 @@ import numpy as np
 import matplotlib as mpl
 from matplotlib import pyplot as plt, ticker, dates as mdates
 from mpl_toolkits.axes_grid1 import make_axes_locatable
+from scipy import stats
 
 from mintpy.objects.coord import coordinate
 from mintpy.objects.colors import ColormapExt
@@ -788,6 +789,7 @@ def plot_perp_baseline_hist(ax, dateList, pbaseList, p_dict={}, dateList_drop=[]
 
     return ax
 
+
 def plot_rotate_diag_coherence_matrix(ax, coh_list, date12_list, date12_list_drop=[],
                                       rotate_deg=-45., cmap='RdBu', disp_half=False, disp_min=0.2):
     """Plot Rotated Coherence Matrix, suitable for Sentinel-1 data with sequential network"""
@@ -927,9 +929,173 @@ def plot_coherence_matrix(ax, date12List, cohList, date12List_drop=[], p_dict={}
     return ax, coh_mat, im
 
 
+def plot_num_triplet_with_nonzero_integer_ambiguity(fname, display=False, font_size=12, fig_size=[9,3]):
+    """Plot the histogram for the number of triplets with non-zero integer ambiguity.
 
+    Fig. 3d-e in Yunjun et al. (2019, CAGEO).
+
+    Parameters: fname - str, path to the numTriNonzeroIntAmbiguity.h5 file.
+    """
+
+    # read data
+    data, atr = readfile.read(fname)
+    vmax = int(np.nanmax(data))
+
+    # plot
+    fig, axs = plt.subplots(nrows=1, ncols=2, figsize=fig_size)
+
+    # subplot 1 - map
+    ax = axs[0]
+    im = ax.imshow(data, cmap='RdBu_r', interpolation='nearest')
+
+    # reference point
+    if all(key in atr.keys() for key in ['REF_Y','REF_X']):
+        ax.plot(int(atr['REF_X']), int(atr['REF_Y']), 's', color='white', ms=3)
+
+    # format
+    auto_flip_direction(atr, ax=ax, print_msg=False)
+    fig.colorbar(im, ax=ax)
+    ax.set_title(r'$T_{int}$', fontsize=font_size)
+
+    # subplot 2 - histogram
+    ax = axs[1]
+    ax.hist(data[~np.isnan(data)].flatten(), range=(0, vmax), log=True, bins=vmax)
+
+    # axis format
+    ax.set_xlabel(r'# of triplets w non-zero int ambiguity $T_{int}$', fontsize=font_size)
+    ax.set_ylabel('# of pixels', fontsize=font_size)
+    ax.xaxis.set_minor_locator(ticker.AutoMinorLocator())
+    ax.yaxis.set_major_locator(ticker.LogLocator(base=10.0, numticks=15))
+    ax.yaxis.set_minor_locator(ticker.LogLocator(base=10.0, numticks=15,
+                                                 subs=(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9)))
+    ax.yaxis.set_minor_formatter(ticker.NullFormatter())
+
+    for ax in axs:
+        ax.tick_params(which='both', direction='in', labelsize=font_size,
+                       bottom=True, top=True, left=True, right=True)
+
+    fig.tight_layout()
+
+    # output
+    out_fig = '{}.png'.format(os.path.splitext(fname)[0])
+    print('plot and save figure to file', out_fig)
+    fig.savefig(out_fig, bbox_inches='tight', transparent=True, dpi=300)
+
+    if display:
+        plt.show()
+    else:
+        plt.close(fig)
+
+    return
+
+
+def plot_timeseries_rms(rms_file, cutoff=3, out_fig=None, disp_fig=True,
+                        fig_size=[5, 3], font_size=12, tick_year_num=1, legend_loc='best',
+                        disp_legend=True, disp_side_plot=True, disp_thres_text=False,
+                        ylabel='Residual phase RMS [mm]'):
+    """ Bar plot for the phase residual RMS time series.
+
+    Parameters: rms_file - str, path to the time series RMS text file
+                           Generated by utils1.get_residual_rms().
+                cutoff   - float, cutoff value of MAD outlier detection
+                fig_size - list of 2 float, figure size in inch
+    """
+
+    # read date / RMS info from text file
+    fc = np.loadtxt(rms_file, dtype=bytes).astype(str)
+    rms_list = fc[:, 1].astype(np.float32) * 1000.
+    date_list = list(fc[:, 0])
+
+    dates, datevector = ptime.date_list2vector(date_list)
+    dates = np.array(dates)
+    try:
+        bar_width = min(ut0.most_common(np.diff(dates).tolist(), k=2))*3/4
+    except:
+        bar_width = np.min(np.diff(dates).tolist())*3/4
+    rms = np.array(rms_list)
+
+    # Plot all dates
+    fig, ax = plt.subplots(figsize=fig_size)
+    ax.bar(dates, rms, bar_width.days, color='C0')
+
+    # Plot reference date
+    ref_idx = np.argmin(rms)
+    ax.bar(dates[ref_idx], rms[ref_idx], bar_width.days, color='C1', label='Reference date')
+
+    # Plot exclude dates
+    rms_threshold = ut0.median_abs_deviation_threshold(rms, center=0., cutoff=cutoff)
+    ex_idx = rms > rms_threshold
+    if not np.all(ex_idx==False):
+        ax.bar(dates[ex_idx], rms[ex_idx], bar_width.days, color='darkgray', label='Exclude date')
+
+    # Plot rms_threshold line
+    (ax, xmin, xmax) = auto_adjust_xaxis_date(ax, datevector, font_size, every_year=tick_year_num)
+    ax.plot(np.array([xmin, xmax]), np.array([rms_threshold, rms_threshold]), '--k',
+            label='Median Abs Dev * {}'.format(cutoff))
+
+    # axis format
+    ax = auto_adjust_yaxis(ax, np.append(rms, rms_threshold), font_size, ymin=0.0)
+    #ax.set_xlabel('Time [years]', fontsize=font_size)
+    ax.set_ylabel(ylabel, fontsize=font_size)
+    ax.tick_params(which='both', direction='in', labelsize=font_size,
+                   bottom=True, top=True, left=True, right=True)
+
+    # 2nd axes for circles
+    if disp_side_plot:
+        divider = make_axes_locatable(ax)
+        ax2 = divider.append_axes("right", "10%", pad="2%")
+        ax2.plot(np.ones(rms.shape, np.float32) * 0.5, rms, 'o', mfc='none', color='C0')
+        ax2.plot(np.ones(rms.shape, np.float32)[ref_idx] * 0.5, rms[ref_idx], 'o', mfc='none', color='C1')
+        if not np.all(ex_idx==False):
+            ax2.plot(np.ones(rms.shape, np.float32)[ex_idx] * 0.5, rms[ex_idx], 'o', mfc='none', color='darkgray')
+        ax2.plot(np.array([0, 1]), np.array([rms_threshold, rms_threshold]), '--k')
+
+        ax2.set_ylim(ax.get_ylim())
+        ax2.set_xlim([0, 1])
+        ax2.tick_params(which='both', direction='in', labelsize=font_size,
+                        bottom=True, top=True, left=True, right=True)
+        ax2.get_xaxis().set_ticks([])
+        ax2.get_yaxis().set_ticklabels([])
+
+    if disp_legend:
+        ax.legend(loc=legend_loc, frameon=False, fontsize=font_size)
+
+    # rms_threshold text
+    if disp_thres_text:
+        ymin, ymax = ax.get_ylim()
+        yoff = (ymax - ymin) * 0.1
+        if (rms_threshold - ymin) > 0.5 * (ymax - ymin):
+            yoff *= -1.
+        ax.annotate('Median Abs Dev * {}'.format(cutoff),
+                    xy=(xmin + (xmax-xmin)*0.05, rms_threshold + yoff ),
+                    color='k', xycoords='data', fontsize=font_size)
+
+    # figure output
+    if out_fig:
+        print('save figure to file:', out_fig)
+        fig.savefig(out_fig, bbox_inches='tight', transparent=True)
+
+    if disp_fig:
+        plt.show()
+    else:
+        plt.close()
+
+    return
+
+
+
+###############################################  GNSS  ###############################################
 
 def plot_gps(ax, SNWE, inps, metadata=dict(), print_msg=True):
+    """Plot GNSS as scatters on top of the input matplotlib.axes.
+
+    Parameters: ax       - matplotlib.axes object
+                SNWE     - tuple of 4 float, for south, north, west and east
+                inps     - Namespace object, from view.py
+                metadata - dict, mintpy metadata
+    Returns:    ax       - matplotlib.axes object
+    """
+
     from mintpy.objects import gps
     vprint = print if print_msg else lambda *args, **kwargs: None
 
@@ -1045,6 +1211,143 @@ def plot_gps(ax, SNWE, inps, metadata=dict(), print_msg=True):
             ax.annotate(site_name, xy=(lon, lat), fontsize=inps.font_size)
 
     return ax
+
+
+def plot_insar_vs_gps_scatter(vel_file, csv_file='gps_enu2los.csv', msk_file=None, ref_gps_site=None, cutoff=5,
+                              fig_size=[4, 4], xname='InSAR', vlim=None, ex_gps_sites=[], display=True):
+    """Scatter plot to compare the velocities between SAR/InSAR and GPS.
+
+    Parameters: vel_file     - str, path of InSAR LOS velocity HDF5 file.
+                csv_file     - str, path of GNSS CSV file, generated after running view.py --gps-comp
+                msk_file     - str, path of InSAR mask file.
+                ref_gps_site - str, reference GNSS site name
+                cutoff       - float, threshold in terms of med abs dev (MAD) for outlier detection
+                xname        - str, xaxis label
+                vlim         - list of 2 float, display value range in the unit of cm/yr
+                               Default is None to grab from data
+                               If set, the range will be used to prune the SAR and GPS observations
+                ex_gps_sites - list of str, exclude GNSS sites for analysis and plotting.
+    Returns:    sites        - list of str, GNSS site names used for comparison
+                insar_obs    - 1D np.ndarray in float32, InSAR velocity in cm/yr
+                gps_obs      - 1D np.ndarray in float32, GNSS  velocity in cm/yr
+    Example:
+        from mintpy.utils import plot as pp
+        csv_file = os.path.join(work_dir, 'geo/gps_enu2los.csv')
+        vel_file = os.path.join(work_dir, 'geo/geo_velocity.h5')
+        msk_file = os.path.join(work_dir, 'geo/geo_maskTempCoh.h5')
+        pp.plot_insar_vs_gps_scatter(vel_file, ref_gps_site='CACT', csv_file=csv_file, msk_file=msk_file, vlim=[-2.5, 2])
+    """
+
+    disp_unit = 'cm/yr'
+    unit_fac = 100.
+
+    # read GPS velocity from CSV file (generated by gps.get_gps_los_obs())
+    col_names = ['Site', 'Lon', 'Lat', 'Displacement', 'Velocity']
+    num_col = len(col_names)
+    col_types = ['U10'] + ['f8'] * (num_col - 1)
+
+    print('read GPS velocity from file: {}'.format(csv_file))
+    fc = np.genfromtxt(csv_file, dtype=col_types, delimiter=',', names=True)
+    sites = fc['Site']
+    lats = fc['Lat']
+    lons = fc['Lon']
+    gps_obs = fc[col_names[-1]] * unit_fac
+
+    if ex_gps_sites:
+        ex_flag = np.array([x in ex_gps_sites for x in sites], dtype=np.bool_)
+        if np.sum(ex_flag) > 0:
+            sites = sites[~ex_flag]
+            lats = lats[~ex_flag]
+            lons = lons[~ex_flag]
+            gps_obs = gps_obs[~ex_flag]
+
+    # read InSAR velocity
+    print('read InSAR velocity from file: {}'.format(vel_file))
+    atr = readfile.read_attribute(vel_file)
+    length, width = int(atr['LENGTH']), int(atr['WIDTH'])
+    ys, xs = coordinate(atr).geo2radar(lats, lons)[:2]
+
+    msk = readfile.read(msk_file)[0] if msk_file else np.ones((length, width), dtype=np.bool_)
+
+    num_site = sites.size
+    insar_obs = np.zeros(num_site, dtype=np.float32) * np.nan
+    prog_bar = ptime.progressBar(maxValue=num_site)
+    for i in range(num_site):
+        x, y = xs[i], ys[i]
+        if (0 <= x < width) and (0 <= y < length) and msk[y, x]:
+            box = (x, y, x+1, y+1)
+            insar_obs[i] = readfile.read(vel_file, datasetName='velocity', box=box)[0] * unit_fac
+        prog_bar.update(i+1, suffix='{}/{} {}'.format(i+1, num_site, sites[i]))
+    prog_bar.close()
+
+    off_med = np.nanmedian(insar_obs - gps_obs)
+    print(f'median offset between InSAR and GPS [before common referencing]: {off_med:.2f} cm/year')
+
+    # reference site
+    if ref_gps_site:
+        print(f'referencing both InSAR and GPS data to site: {ref_gps_site}')
+        ref_ind = sites.tolist().index(ref_gps_site)
+        gps_obs -= gps_obs[ref_ind]
+        insar_obs -= insar_obs[ref_ind]
+
+    # remove NaN value
+    print('removing sites with NaN values in GPS or {}'.format(xname))
+    flag = np.multiply(~np.isnan(insar_obs), ~np.isnan(gps_obs))
+    if vlim is not None:
+        print('pruning sites with value range: {} {}'.format(vlim, disp_unit))
+        flag *= gps_obs >= vlim[0]
+        flag *= gps_obs <= vlim[1]
+        flag *= insar_obs >= vlim[0]
+        flag *= insar_obs <= vlim[1]
+
+    gps_obs = gps_obs[flag]
+    insar_obs = insar_obs[flag]
+    sites = sites[flag]
+
+    # stats
+    print('GPS   min/max: {:.2f} / {:.2f}'.format(np.nanmin(gps_obs), np.nanmax(gps_obs)))
+    print('InSAR min/max: {:.2f} / {:.2f}'.format(np.nanmin(insar_obs), np.nanmax(insar_obs)))
+
+    rmse = np.sqrt(np.sum((insar_obs - gps_obs)**2) / (gps_obs.size - 1))
+    r2 = stats.linregress(insar_obs, gps_obs)[2]
+    print('RMSE = {:.2f} {}'.format(rmse, disp_unit))
+    print('R^2 = {:.2f}'.format(r2))
+
+    # preliminary outlier detection
+    diff_mad = ut0.median_abs_deviation(abs(insar_obs - gps_obs), center=0)
+    print(f'Preliminary outliers detection: abs(InSAR - GNSS) > med abs dev ({diff_mad:.2f}) * {cutoff}')
+    print('Site:  InSAR  GNSS')
+    for site_name, insar_val, gps_val in zip(sites, insar_obs, gps_obs):
+        if abs(insar_val - gps_val) > diff_mad * cutoff:
+            print(f'{site_name:s}: {insar_val:5.1f}, {gps_val:5.1f}  {disp_unit}')
+
+    # plot
+    if display:
+        plt.rcParams.update({'font.size': 12})
+        if vlim is None:
+            vlim = [np.min(insar_obs), np.max(insar_obs)]
+            vbuffer = (vlim[1] - vlim[0]) * 0.2
+            vlim = [vlim[0] - vbuffer, vlim[1] + vbuffer]
+
+        fig, ax = plt.subplots(figsize=fig_size)
+        ax.plot((vlim[0], vlim[1]), (vlim[0], vlim[1]), 'k--')
+        ax.plot(insar_obs, gps_obs, '.', ms=15)
+
+        # axis format
+        ax.set_xlim(vlim)
+        ax.set_ylim(vlim)
+        ax.set_xlabel(f'{xname} [{disp_unit}]')
+        ax.set_ylabel(f'GNSS [{disp_unit}]')
+        ax.set_aspect('equal', 'box')
+        fig.tight_layout()
+
+        # output
+        out_fig = '{}_vs_gps_scatter.pdf'.format(xname.lower())
+        plt.savefig(out_fig, bbox_inches='tight', transparent=True, dpi=300)
+        print('save figure to file', out_fig)
+        plt.show()
+
+    return sites, insar_obs, gps_obs
 
 
 def plot_colorbar(inps, im, cax):
